Fix bug parsing system metadata with missing replication attributes.
Added error checking to parseSystemMetadata() so that it first checks for the existince of the `replicationAllowed` and `numberReplicas` attributes before it tries to access them.  Previously, and error was thrown when they were missing.  While it makes no sense to be missing these attributes when `replicationPolicy` is provided, they are optional in the schema.  Now, if they are missing, the parser sets them to sensibel default (replicationAllowed=false, numberReplicas=0). Closes #63.