Refactor DataPackage to not use anonymous methods.
The S4 Methods definitions were previously anonymous, which prevented the R debugger to find the code when entering in to debug a call. Refactored all methods to create a private S3 function that implements each methods, and then reference that in `setMethod`. The generics are still anonymous which could still cause issues.